requirements.txt
chardet
mido
Pillow